Refactor project structure and update imports
Moved components and pages into dedicated 'components' and 'pages' directories, updated all relevant import paths, and removed unused test, config, and asset files. This improves code organization and maintainability.
