+++ src/pages/Records.js
@@ -8,15 +8,15 @@ import {
 } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
 
-import Header from './header';
-import QuickActionButton from './QuickActionButton';
+import Header from '../components/header';
+import QuickActionButton from '../components/QuickActionButton';
 import AddPatientRecord from './add-record';
-import DataTable from './DataTable';
-import SearchBar from './SearchBar';
-import FilterComponent, { FilterButton, FilterContent } from './FilterComponent';
-import SortableHeader, { sortData } from './SortableHeader';
-import DualSortableHeader, { sortDualData } from './DualSortableHeader';
-import Pagination from './Pagination';
+import DataTable from '../components/DataTable';
+import SearchBar from '../components/SearchBar';
+import FilterComponent, { FilterButton, FilterContent } from '../components/FilterComponent';
+import SortableHeader, { sortData } from '../components/SortableHeader';
+import DualSortableHeader, { sortDualData } from '../components/DualSortableHeader';
+import Pagination from '../components/Pagination';
 import ViewRecord from './view-record';
 
 // API Base URL
