+++ src/pages/view-service.js
@@ -16,7 +16,7 @@ import {
 import CloseIcon from '@mui/icons-material/Close';
 import EditIcon from '@mui/icons-material/Edit';
 import SaveIcon from '@mui/icons-material/Save';
-import { API_BASE } from './apiConfig';
+import { API_BASE } from '../apiConfig';
 
 const serviceTypes = [
   'Single Treatment',
